fix(inkling): DCP view-load adapter, uniform dtype for FSDP2
- state_dict_adapter: map HF-module <-> native with the routed-expert transpose
  as a non-contiguous strided view (DTensor/on-device) so DCP loads experts
  in-place; track them in view_loaded_native_keys and skip re-emit in from_hf.
  Plain (full-state) tensors still transpose + EP-slice.
- Store all params in the model dtype (drop fp32 pinning; router bias follows
  model dtype) so FSDP2 sees a uniform dtype per group; sconv/router math still
  upcast to fp32 in compute.

+++ nemo_automodel/components/models/inkling/layers.py
@@ -97,8 +97,9 @@ class InklingGate(nn.Module):
         model_dtype = get_dtype(getattr(config, "torch_dtype", None), torch.bfloat16)
         self.weight = nn.Parameter(torch.empty(self.n_total_experts, config.hidden_size, dtype=model_dtype))
         self.global_scale = nn.Parameter(torch.ones(1, dtype=model_dtype))
-        # Kept in fp32: small quantization errors in bf16 flip expert selection.
-        self.e_score_correction_bias = nn.Parameter(torch.empty(self.num_experts, dtype=torch.float32))
+        # Stored in the model dtype (uniform dtype keeps FSDP2 happy); selection math
+        # upcasts to fp32 via ``gate_precision`` in forward.
+        self.e_score_correction_bias = nn.Parameter(torch.empty(self.num_experts, dtype=model_dtype))
 
     def forward(self, hidden_states: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
         """Route tokens to experts.
